pyproject.toml
[project]
name = "acgn-character-skill"
version = "0.1.0"
description = "OCR-first dialogue extraction and character skill generation for ACGN works."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Pillow>=10.0",
  "numpy>=1.24",
  "av>=12.0",
  "PyYAML>=6.0",
  "tqdm>=4.66",
  "paddlepaddle>=2.6",
  "paddleocr>=2.8",
  "opencv-python>=4.8",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
ocr-fallback = [
  "easyocr>=1.7",
  "rapidocr-onnxruntime>=1.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
